+++ src/main/java/net/glowstone/generator/ground/RockyMountainGroundGenerator.java
@@ -6,7 +6,7 @@ import org.bukkit.Material;
 
 public class RockyMountainGroundGenerator extends GroundGenerator {
     @Override
-    public void generateTerrainColumn(byte[][] buf, Random random, int x, int z, int seaLevel, double surfaceNoise) {
+    public void generateTerrainColumn(short[][] buf, Random random, int x, int z, int seaLevel, double surfaceNoise) {
         if (surfaceNoise > 1.0D) {
             setTopMaterial(Material.STONE);
             setGroundMaterial(Material.STONE);
